Move MatchingContext to a new module
